requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
google-genai
jinja2
python-multipart
